Audit crate exports, add prelude
