+++ event-store/src/internals/test_helpers.rs
